refactor: move helmcharttesting metrics to metrics package, unify imports
On-behalf-of: @SAP krzysztof.zagorski@sap.com
