+++ internal/metrics/team_metrics.go
@@ -5,6 +5,7 @@ package metrics
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
+	crmetrics "sigs.k8s.io/controller-runtime/pkg/metrics"
 
 	greenhousev1alpha1 "github.com/cloudoperators/greenhouse/api/v1alpha1"
 )
@@ -20,7 +21,7 @@ var (
 )
 
 func init() {
-	prometheus.MustRegister(membersCountMetric)
+	crmetrics.Registry.MustRegister(membersCountMetric)
 }
 
 func UpdateTeamMembersCountMetric(team *greenhousev1alpha1.Team, membersCount int) {
